refactor(tests): merge validation tests (Phase 3)
Merge 2 validation test files into single test_validation.py:
- test_validation_guard.py (302 LOC) → Section 1: Validation Guards
- test_revalidation.py (182 LOC) → Section 2: Revalidation Lifecycle

Result: test_validation.py (509 LOC)
- 23 tests preserved (12 standalone + 2 test classes)
- Organized into 2 sections with clear dividers
- Comprehensive docstring explaining guard pattern and revalidation
- All 335 tests still passing
- Consolidated duplicate imports

Test File Summary (14→8 files):
✓ Phase 0: Fixtures extraction (conftest.py)
✓ Phase 1: Parser tests (3→1: test_parser.py)
✓ Phase 2: Integration tests (4→1: test_lndl.py)
✓ Phase 3: Validation tests (2→1: test_validation.py)

Remaining files (5):
- test_lexer.py (lexer unit tests)
- test_resolver_scalar.py (resolver edge cases)
- test_types.py (type system tests)
- test_prompt.py (prompt utility test)
- conftest.py (shared fixtures)
